Sumi's Commit
Everything should be working, there is an sql file along with this for the staff members details, you can see the login info there for staff members and stuff.
If you load StaffLogin.FXML in main, the login should work and it should display the staff (managers) name based on their id and user. There is also a logout feature.
If you load ManageStaff.FXML in main it should show the staff table list along with insert, edit and delete, those should all work and the changes should be committed in the database for persistence. 
My only issue is not being able to load due to the table methods not pointing and being filled up correctly, once that is sorted everything should be done on my end (i think).
This is probably a very messy upload, apologies.

Contents: 
.JAVA: Manager, ManageStaff, Controller
.FXML: Manager, ManageStaff, StaffLogin
.SQL: staffdatabase (or managestaff)
^ If you just want the direct files, just ctrl+f these names i guess.